pyproject.toml
[tool.poetry]
name = "spectrum-retention"
version = "0.1.0"
description = ""
authors = ["matheus_campos <matheus@mobi2buy.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.1.3"
boto3 = "^1.16.0"
python-decouple = "^3.3"
hydra-core = "0.11.3"
pyarrow = "^2.0.0"
matplotlib = "^3.3.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
